+++ sanic/response/__init__.py
@@ -1,4 +1,4 @@
-from .functions import (
+from .convenience import (
     empty,
     file,
     file_stream,
@@ -9,7 +9,7 @@ from .functions import (
     text,
     validate_file,
 )
-from .models import (
+from .types import (
     BaseHTTPResponse,
     HTTPResponse,
     JSONResponse,
