+++ sanic/response/types.py
